fix: ISO8601 duration wouldn't parse for valid fractional values (#199)
* fix: ISO8601 duration won't parse for valid fractional values

Previously, the library utilized a parsing library for ISO8601 durations that parsed durations into integers only. This caused issues with durations containing fractional values, such as "P0.5S", which are valid ISO8601 strings but failed to parse due to the limitation of the previous library.
In this commit, I've replaced the outdated parsing library with a more robust alternative. The new library properly handles fractional values, ensuring accurate parsing of ISO8601 duration strings, including cases like "P0.5S".
This update enhances the functionality and reliability of the library by accommodating a broader range of ISO8601 duration formats. Additionally, it ensures compatibility with modern standards and improves the overall usability of the library.

Signed-off-by: Kshitiz Agrawal <kshitizagrawal@outlook.com>

* fix: resolve failing test cases

Signed-off-by: Kshitiz Agrawal <kshitizagrawal@outlook.com>

* feat: add test case for fractional ISO duration

Signed-off-by: Kshitiz Agrawal <kshitizagrawal@outlook.com>

---------

Signed-off-by: Kshitiz Agrawal <kshitizagrawal@outlook.com>